# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "B4 network packet processor REST API",
        "title": "B4 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/discovery/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Add discovery preset as a new set",
                "parameters": [
                    {
                        "description": "Set configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.SetConfig"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/discovery/cache/clear": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Clear discovery cache",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/discovery/cancel/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Cancel discovery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Suite ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/discovery/current": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Get current running discovery",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/discovery/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Get discovery history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
        "/discovery/history/clear": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Clear discovery history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/discovery/history/{domain}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Delete discovery history entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Domain name",
                        "name": "domain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/discovery/similar": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Find sets with similar configuration",
                "parameters": [
                    {
                        "description": "Set to compare",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.SetConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
        "/discovery/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Start domain discovery",
                "parameters": [
                    {
                        "description": "Discovery request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DiscoveryRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/handler.DiscoveryResponse"
                        }
                    }
                }
            }
        },
        "/discovery/status/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Get discovery status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Suite ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "config.DNSConfig": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "fragment_query": {
                    "type": "boolean"
                },
                "target_dns": {
                    "type": "string"
                }
            }
        },
...
        "config.DuplicateConfig": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "Number of packet copies to send (original is dropped)",
                    "type": "integer"
                },
                "enabled": {
                    "type": "boolean"
                }
            }
        },
        "config.FakingConfig": {
            "type": "object",
            "properties": {
                "custom_payload": {
                    "type": "string"
                },
                "payload_file": {
                    "type": "string"
                },
                "seq_offset": {
                    "type": "integer"
                },
                "sni": {
                    "type": "boolean"
                },
                "sni_mutation": {
                    "$ref": "#/definitions/config.SNIMutationConfig"
                },
                "sni_seq_length": {
                    "type": "integer"
                },
                "sni_type": {
                    "type": "integer"
                },
                "strategy": {
                    "type": "string"
                },
                "tcp_md5": {
                    "description": "Enable TCP MD5 option insertion",
                    "type": "boolean"
                },
                "timestamp_decrease": {
                    "description": "Amount to decrease TCP timestamp option",
                    "type": "integer"
                },
                "tls_mod": {
                    "description": "e.g. [\"rnd\", \"dupsid\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ttl": {
                    "type": "integer"
                }
            }
        },
        "config.FragmentationConfig": {
            "type": "object",
            "properties": {
                "combo": {
                    "$ref": "#/definitions/config.ComboFragConfig"
                },
                "disorder": {
                    "$ref": "#/definitions/config.DisorderFragConfig"
                },
                "middle_sni": {
                    "type": "boolean"
                },
                "oob_char": {
                    "description": "Character for OOB data",
                    "type": "integer"
                },
                "oob_position": {
                    "description": "Position for OOB (0=disabled)",
                    "type": "integer"
                },
                "oob_position_max": {
                    "description": "max for randomization (0 = use fixed)",
                    "type": "integer"
                },
                "reverse_order": {
                    "type": "boolean"
                },
                "seq_overlap_pattern": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sni_position": {
                    "type": "integer"
                },
                "sni_position_max": {
                    "description": "max for randomization (0 = use fixed)",
                    "type": "integer"
                },
                "strategy": {
                    "description": "Values: \"tcp\", \"ip\", \"oob\", \"tls\", \"disorder\",  \"extsplit\", \"firstbyte\", \"combo\", \"none\"",
                    "type": "string"
                },
                "strategy_pool": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tlsrec_pos": {
                    "description": "where to split TLS record",
                    "type": "integer"
                },
                "tlsrec_pos_max": {
                    "description": "max for randomization (0 = use fixed)",
                    "type": "integer"
                }
            }
        },
...
        "config.RoutingConfig": {
            "type": "object",
            "properties": {
                "egress_interface": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "fwmark": {
                    "type": "integer"
                },
                "ip_ttl_seconds": {
                    "type": "integer"
                },
                "source_interfaces": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "table": {
                    "type": "integer"
                }
            }
        },
        "config.SNIMutationConfig": {
            "type": "object",
            "properties": {
                "fake_ext_count": {
                    "type": "integer"
                },
                "fake_snis": {
                    "description": "Additional SNIs to inject",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "grease_count": {
                    "type": "integer"
                },
                "mode": {
                    "description": "\"off\", \"duplicate\", \"grease\", \"padding\", \"reorder\", \"full\"",
                    "type": "string"
                },
                "padding_size": {
                    "type": "integer"
                }
            }
        },
        "config.SetConfig": {
            "type": "object",
            "properties": {
                "dns": {
                    "$ref": "#/definitions/config.DNSConfig"
                },
                "enabled": {
                    "type": "boolean"
                },
                "faking": {
                    "$ref": "#/definitions/config.FakingConfig"
                },
                "fragmentation": {
                    "$ref": "#/definitions/config.FragmentationConfig"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "routing": {
                    "$ref": "#/definitions/config.RoutingConfig"
                },
                "targets": {
                    "$ref": "#/definitions/config.TargetsConfig"
                },
                "tcp": {
                    "$ref": "#/definitions/config.TCPConfig"
                },
                "udp": {
                    "$ref": "#/definitions/config.UDPConfig"
                }
            }
        },
...
        "config.TCPConfig": {
            "type": "object",
            "properties": {
                "conn_bytes_limit": {
                    "type": "integer"
                },
                "desync": {
                    "$ref": "#/definitions/config.DesyncConfig"
                },
                "dport_filter": {
                    "description": "comma separated list of ports and port ranges, e.g. \"80,443,5222\"",
                    "type": "string"
                },
                "drop_sack": {
                    "type": "boolean"
                },
                "duplicate": {
                    "$ref": "#/definitions/config.DuplicateConfig"
                },
                "incoming": {
                    "$ref": "#/definitions/config.IncomingConfig"
                },
                "seg2delay": {
                    "type": "integer"
                },
                "seg2delay_max": {
                    "type": "integer"
                },
                "syn_fake": {
                    "type": "boolean"
                },
                "syn_fake_len": {
                    "type": "integer"
                },
                "syn_ttl": {
                    "type": "integer"
                },
                "win": {
                    "$ref": "#/definitions/config.WinConfig"
                }
            }
        },
...
        "config.TargetsConfig": {
            "type": "object",
            "properties": {
                "geoip_categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "geosite_categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ip": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sni_domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "source_devices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tls": {
                    "description": "\"1.2\", \"1.3\", or \"\" (match any)",
                    "type": "string"
                }
            }
        },
        "config.UDPConfig": {
            "type": "object",
            "properties": {
                "conn_bytes_limit": {
                    "type": "integer"
                },
                "dport_filter": {
                    "description": "can be a comma separated list of ports and port ranges, e.g. \"80,443,1000-2000\"",
                    "type": "string"
                },
                "fake_len": {
                    "type": "integer"
                },
                "fake_seq_length": {
                    "type": "integer"
                },
                "faking_strategy": {
                    "type": "string"
                },
                "filter_quic": {
                    "type": "string"
                },
                "filter_stun": {
                    "type": "boolean"
                },
                "mode": {
                    "type": "string"
                },
                "seg2delay": {
                    "type": "integer"
                },
                "seg2delay_max": {
                    "type": "integer"
                }
            }
        },
...
        "handler.DiscoveryRequest": {
            "type": "object",
            "properties": {
                "check_url": {
                    "type": "string"
                },
                "check_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "payload_files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "skip_cache": {
                    "type": "boolean"
                },
                "skip_dns": {
                    "type": "boolean"
                },
                "tls_version": {
                    "description": "\"auto\", \"tls12\", \"tls13\"",
                    "type": "string"
                },
                "validation_tries": {
                    "type": "integer"
                }
            }
        },
        "handler.DiscoveryResponse": {
            "type": "object",
            "properties": {
                "check_url": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "estimated_tests": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\" to authorize",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}